pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Task manager"
authors = ["Cvilev Andrey <CvilevAS@mail.ru>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
python-dotenv = "^1.0.0"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
django-bootstrap5 = "^23.3"
django-filter = "^23.5"
rollbar = "^1.0.0"


[tool.poetry.group.dev.dependencies]
selenium = "^4.16.0"
coverage = "^7.3.2"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
